Move variables hooks to hooks directory
